Modify so that only the room creator deletes ...
the room when executes closeAndCleaning

+++ src/resources/js/data_channel.js
@@ -259,7 +259,7 @@ export async function closeAndCleaning() {
     peerConnection.close();
   }
   // Delete room on hangup
-  if (roomId) {
+  if (channel.amICreatedRoom && roomId) {
     const db = firebase.firestore();
     const roomRef = db.collection('rooms').doc(roomId);
     await roomRef.delete();
